Fix relay selector to force blocked state with daita and obfuscation on
